Migration to system class loader implementation
Rather than executing after class loading has already started and then
having to deal with reloading various classes, this is now intended to
be used as a system class loader in its own right, either directly
specified on the command line or pulled in by a stub loader. Thus, the
Bootstrapper class is no longer necessary, and ModClassLoader now
implements main() by dynamically generating a class of the proper name.

This also removes Javassist from the build, as I'm settling on ASM.